escreve instruções e palavras

+++ utils.h
@@ -7,7 +7,7 @@ int FileIsNull(FILE *input);
 int removerX(char *str);
 void removerNewLine(char *str);
 unsigned int strToBin(const char *strBinario);
-void printBinary(int n);
+void printBinary(long long n);
 void line (char *symbols, int length);
 
 #endif
